// herevemarket-app-main/src/styles/home.styles.d.ts
import { ViewStyle, TextStyle, ImageStyle } from "react-native";

declare module "../styles/home.styles" {
  export const styles: { [key: string]: ViewStyle | TextStyle | ImageStyle };
  export const styles: {
    container: ViewStyle;
    header: ViewStyle;
    title: TextStyle;
    categoryScroll: ViewStyle;
    categoryCard: ViewStyle;
    categoryCardActive: ViewStyle;
    categoryText: TextStyle;
    categoryTextActive: TextStyle;
    brandScroll: ViewStyle;
    brandChip: ViewStyle;
    brandImage: ImageStyle;
    bannerWrapper: ViewStyle;
    banner: ViewStyle;
    bannerImage: ImageStyle;
    bannerText: TextStyle;
    bannerDots: ViewStyle;
    bannerDot: ViewStyle;
    bannerDotActive: ViewStyle;
    products: ViewStyle;
    sectionTitle: TextStyle;
    productCard: ViewStyle;
    productName: TextStyle;
    productPrice: TextStyle;
    productImage: ImageStyle;
    productInfo: ViewStyle;
    productFooter: ViewStyle;
    addButton: ViewStyle;
    addButtonText: TextStyle;
    counter: ViewStyle;
    counterButton: ViewStyle;
    counterText: TextStyle;
    counterValue: TextStyle;
    logo: ImageStyle;
    floatingCart: ViewStyle;
    floatingCartIcon: TextStyle;
    cartBadge: ViewStyle;
    cartBadgeRow: ViewStyle;
    cartBadgeText: TextStyle;
    cartEtaPill: ViewStyle;
    cartEtaText: TextStyle;
    cartHeroTitle: TextStyle;
    cartHeroText: TextStyle;
    cartSectionCard: ViewStyle;
    cartSectionTitle: TextStyle;
    cartInfoPill: ViewStyle;
    cartInfoPillText: TextStyle;
    cartHero: ViewStyle;
    cartHeroBackdrop: ViewStyle;
    cartContainer: ViewStyle;
    cartHeader: ViewStyle;
    cartBackButton: ViewStyle;
    cartBackText: TextStyle;
    cartTitle: TextStyle;
    cartSubtitle: TextStyle;
    cartList: ViewStyle;
    cartItem: ViewStyle;
    cartItemImage: ImageStyle;
    cartItemInfo: ViewStyle;
    cartItemName: TextStyle;
    cartItemMeta: ViewStyle;
    cartItemPrice: TextStyle;
    cartQuantityBadge: ViewStyle;
    cartQuantityText: TextStyle;
    cartItemTotal: TextStyle;
    cartTotalCard: ViewStyle;
    cartTotalRow: ViewStyle;
    cartTotalLabel: TextStyle;
    cartTotalValue: TextStyle;
    cartDivider: ViewStyle;
    cartGrandTotalLabel: TextStyle;
    cartGrandTotalValue: TextStyle;
    checkoutButton: ViewStyle;
    checkoutButtonText: TextStyle;
    cartEmpty: ViewStyle;
    page: ViewStyle;
    topSection: ViewStyle;
    categoryBar: ViewStyle;
    categoryRow: ViewStyle;
    contentContainer: ViewStyle;
    contentScroll: ViewStyle;
    cartEmptyIcon: TextStyle;
    cartEmptyTitle: TextStyle;
    cartEmptyText: TextStyle;
    cartBackButtonAlt: ViewStyle;
    cartBackTextAlt: TextStyle;
    cartBadgeText: TextStyle;
    legalSection: ViewStyle;
    legalLinksRow: ViewStyle;
    legalLinksColumn: ViewStyle;
    legalLink: ViewStyle;
    legalLinkText: TextStyle;
    legalCaption: TextStyle;
    legalLogosRow: ViewStyle;
    legalLogo: ImageStyle;
    legalLogoWide: ImageStyle;
  };
}